Fix concurrent download issue by using unique temporary directories

+++ backend/src/usecase/downloads/remove_files.py
@@ -1,11 +1,13 @@
 import os
 import shutil
 
-from src.usecase.downloads.tmp_path import TMP_DIR_PATH
-from src.utils.path import generate_parent_abs_path
 
-
-def remove_files():
-    tmp_dir_path = generate_parent_abs_path(TMP_DIR_PATH)
-    shutil.rmtree(tmp_dir_path)
-    os.mkdir(tmp_dir_path)
+def remove_files(tmp_dir_path):
+    """
+    Remove the temporary directory and all its contents.
+    
+    Args:
+        tmp_dir_path: Path to the temporary directory to remove
+    """
+    if os.path.exists(tmp_dir_path):
+        shutil.rmtree(tmp_dir_path)
